+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRCVFVBGNQW41Z77WM7YPEDH"
+        "01KRDF8KJQ5MT5Z19746PE9WW4"
     )
     r"""The unique identifier of the data source"""