+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRCVFVBJ12ZE17CCG9XA5CPT"
+        "01KRDF8KJT43ZEA2N38TDQDH08"
     )
     r"""The unique identifier of the data source"""