+++ packages/orq-rc/RELEASES.md
@@ -7779,3 +7779,13 @@ Based on:
 - [python v4.9.0-rc.45] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.45] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.45 - packages/orq-rc
+
+## 2026-05-12 06:54:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.46] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.46] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.46 - packages/orq-rc